+++ src/gui/main_box/menu_box/init_menu_box.c
@@ -59,7 +59,7 @@ static int	init_menu_gui_box_children(t_engine *minirt, t_gui_box *gui_box)
 static int	init_camera_icon_box(t_engine *minirt, t_gui_box *gui_box,
 				t_gui_box *parent)
 {
-	const int	box_width = round(((double)parent->size.x - ICON_BOX_SEPARATOR * 4)
+	const int	box_width = roundf(((float)parent->size.x - ICON_BOX_SEPARATOR * 4)
 								   / 3.0);
 
 	*gui_box = create_t_gui_box(minirt, NULL, \
@@ -89,7 +89,7 @@ static int	init_camera_icon_box(t_engine *minirt, t_gui_box *gui_box,
 static int	init_saving_icon_box(t_engine *minirt, t_gui_box *gui_box,
 								   t_gui_box *parent)
 {
-	const int	box_width = round(((double)parent->size.x - ICON_BOX_SEPARATOR * 4)
+	const int	box_width = roundf(((float)parent->size.x - ICON_BOX_SEPARATOR * 4)
 								   / 3.0);
 
 	*gui_box = create_t_gui_box(minirt, NULL, \
@@ -119,7 +119,7 @@ static int	init_saving_icon_box(t_engine *minirt, t_gui_box *gui_box,
 static int	init_menu_settings_icon_box(t_engine *minirt, t_gui_box *gui_box,
 										  t_gui_box *parent)
 {
-	const int	box_width = round(((double)parent->size.x - ICON_BOX_SEPARATOR * 4)
+	const int	box_width = roundf(((float)parent->size.x - ICON_BOX_SEPARATOR * 4)
 			/ 3.0);
 
 	*gui_box = create_t_gui_box(minirt, NULL, \
