+++ src/ray_tracer/camera/movement.c
@@ -12,13 +12,13 @@
 
 #include "ray_tracer/camera.h"
 
-void	camera_move_forward(t_camera *camera, double distance)
+void	camera_move_forward(t_camera *camera, float distance)
 {
 	camera->position = vector3_add(camera->position, vector3_multiply(
 				vector3_unit(camera->direction), distance));
 }
 
-void	camera_move_left(t_camera *camera, double distance)
+void	camera_move_left(t_camera *camera, float distance)
 {
 	t_vector3	right_direction;
 
@@ -27,14 +27,14 @@ void	camera_move_left(t_camera *camera, double distance)
 				vector3_unit(right_direction), distance));
 }
 
-void	camera_rotate_left(t_camera *camera, double degrees)
+void	camera_rotate_left(t_camera *camera, float degrees)
 {
 	camera->direction = vector3_rotate_y(camera->direction, degrees);
 }
 
-void	camera_rotate_up(t_camera *camera, double degrees)
+void	camera_rotate_up(t_camera *camera, float degrees)
 {
-	const double	up = degrees * (1.0 / 90.0);
+	const float	up = degrees * (1.0 / 90.0);
 
 	if (camera->direction.y + up < 1.0 && camera->direction.y + up > -1.0)
 		camera->direction.y += up;
